MBS-10118 Rename RxIdlingRegistry.kt to the Rx3IdlingRegistry.kt add brief motivation for that module as a KDoc
