+++ buildscript/src/main/kotlin/Dependencies.kt
@@ -93,7 +93,7 @@ object Dependencies {
         const val espressoIntents = "androidx.test.espresso:espresso-intents:${Versions.espresso}"
         const val espressoDescendantActions = "com.forkingcode.espresso.contrib:espresso-descendant-actions:1.4.0"
         const val kaspresso = "com.kaspersky.android-components:kaspresso:1.1.0"
-        const val rxIlder = "com.squareup.rx.idler:rx3-idler:0.11.0"
+        const val rx3Ilder = "com.squareup.rx.idler:rx3-idler:0.11.0"
     }
 
     object Test {
