--- subprojects/android-test/rx-idler/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="com.avito.android.test.rxidler" />
