+++ subprojects/android-test/rx3-idler/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.avito.android.test.rx3idler" />
