+++ subprojects/android-test/rx3-idler/src/main/kotlin/com/avito/android/test/rx3idler/Rx3IdlingRegistry.kt
@@ -1,4 +1,4 @@
-package com.avito.android.test.rxidler
+package com.avito.android.test.rx3idler
 
 import androidx.test.espresso.IdlingRegistry
 import androidx.test.espresso.IdlingResource
@@ -9,9 +9,14 @@ import io.reactivex.rxjava3.functions.Supplier
 import io.reactivex.rxjava3.plugins.RxJavaPlugins
 
 /**
- * Motivation https://github.com/square/RxIdler/issues/34
+ * Simple registration of rx [Schedulers] as an [IdlingResource] via [Rx3Idler] doesn't work because:
+ *  - we have tests with infinite [Observable] that causes [IdlingResourceTimeoutException]
+ *
+ *  This class helps us enable and disable [IdlingResource] when we need
+ *
+ * Details https://github.com/square/RxIdler/issues/34
  */
-object RxIdlingRegistry {
+object Rx3IdlingRegistry {
 
     private var initialized = false
     private var registered = false
@@ -23,7 +28,7 @@ object RxIdlingRegistry {
      * read https://github.com/square/RxIdler#usage
      */
     fun initialize() {
-        synchronized(RxIdlingRegistry) {
+        synchronized(Rx3IdlingRegistry) {
             require(!initialized) {
                 "Already have been initialized"
             }
@@ -37,10 +42,10 @@ object RxIdlingRegistry {
      * Enable RX Idling
      * Should be called only after [disable] have already been called
      *
-     * [RxIdlingRegistry] is enabled by default after [initialize]
+     * [Rx3IdlingRegistry] is enabled by default after [initialize]
      */
     fun enable() {
-        synchronized(RxIdlingRegistry) {
+        synchronized(Rx3IdlingRegistry) {
             require(initialized) {
                 "Must be initialized"
             }
@@ -54,14 +59,14 @@ object RxIdlingRegistry {
 
     /**
      * Disable RX Idling
-     * read [RxIdlingRegistry] motivation
+     * read [Rx3IdlingRegistry] motivation
      *
      * you should prefer to use [withDisabled] because you could forgot to enable back
      */
     fun disable() {
-        synchronized(RxIdlingRegistry) {
+        synchronized(Rx3IdlingRegistry) {
             require(initialized) {
-                "Expect initialized = true"
+                "Must be initialized"
             }
             require(registered) {
                 "Expect registered = true"
@@ -75,7 +80,7 @@ object RxIdlingRegistry {
      * Run [action] without RX Idling
      */
     fun withDisabled(action: () -> Unit) {
-        synchronized(RxIdlingRegistry) {
+        synchronized(Rx3IdlingRegistry) {
             val wasRegistered = registered
             if (wasRegistered) {
                 disable()
@@ -96,7 +101,7 @@ object RxIdlingRegistry {
 
     private fun createInitSchedulerHandler(name: String): Function<Supplier<Scheduler>, Scheduler> {
         return Function { scheduler ->
-            synchronized(RxIdlingRegistry) {
+            synchronized(Rx3IdlingRegistry) {
                 val idlingScheduler = Rx3Idler.wrap(scheduler.get(), name)
                 registry.add(idlingScheduler)
                 // This state could be reached if client disabled RxIdlingRegistry before scheduler actually initialized
